security: Prevent arbitrary file deletion in `rm -rf` by validating directories

+++ Lidarr-MusicAutomator.bash
@@ -191,7 +191,7 @@ SearchDeezerAlbums () {
   
             log "$processNumber of $lidarrTotalRecords :: $lidarrAlbumArtistName :: $lidarrAlbumTitle :: $deezerAlbumTitle :: Explicit Lyrics ($deezerExplicitLyrics) :: Match Found!"
 
-            if [ -n "$incompleteDownloadPath" ] && [ "$incompleteDownloadPath" != "/" ] && [ -d "$incompleteDownloadPath" ]; then
+            if [ -n "$incompleteDownloadPath" ] && [ "$incompleteDownloadPath" != "/" ] && [ -d "$incompleteDownloadPath" ] && [ "$(realpath "$incompleteDownloadPath" 2>/dev/null)" != "/" ]; then
                 rm -rf "${incompleteDownloadPath:?}"
             fi
 
@@ -380,7 +380,7 @@ for (( ; ; )); do
         ArlSetup
 
         log "Step - Removing previously downloaded items that failed to import..."
-        if [ -n "$completeDownloadPath" ] && [ "$completeDownloadPath" != "/" ] && [ -d "$completeDownloadPath" ]; then
+        if [ -n "$completeDownloadPath" ] && [ "$completeDownloadPath" != "/" ] && [ -d "$completeDownloadPath" ] && [ "$(realpath "$completeDownloadPath" 2>/dev/null)" != "/" ]; then
             rm -rf "${completeDownloadPath:?}"/*
         fi
